Replace shared pointers with unique pointers Remove unused constructors Cache resources Call  AddGameObject to create a Game Object

+++ Minigin/Utility/Font.cpp
@@ -16,7 +16,7 @@ namespace dae
 
     Font::Font(const std::string& fullPath, unsigned int size) : m_font(nullptr)
     {
-        m_font = TTF_OpenFont(fullPath.c_str(), size);
+        m_font = TTF_OpenFont(fullPath.c_str(), static_cast<int>(size));
         if (m_font == nullptr)
         {
             throw std::runtime_error(std::string("Failed to load font: ") + SDL_GetError());
